bizzer-agents/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0

# Redis
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector DB
chromadb>=0.4.0

# JWT
PyJWT>=2.8.0

# LangChain (optional, for advanced features)
# langchain>=0.1.0
# langchain-community>=0.0.10

# Utils
python-decouple>=3.8
